!36 修复射击模块 Merge pull request !36 from 刘骢/2020

+++ User/component/user_math.c
@@ -87,6 +87,6 @@ inline void CircleAdd(float *origin, float delta, float range) {
  * \param fric_radius_m 摩擦轮半径
  * \return 摩擦轮转速
  */
-inline float CalculateRpm(uint8_t shoot_speed, float fric_radius_m) {
+inline float CalculateRpm(uint16_t shoot_speed, float fric_radius_m) {
   return 60.0f * (float)shoot_speed / (M_2PI * fric_radius_m);
 }
